stock_etl/requirements.txt
akshare==1.17.69
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.8.1
python-json-logger==3.3.0
requests==2.32.3
pytest==8.3.5
